(GH-131) Add Get-BTHeader function
This commit adds the Get-BTHeader function which is used to query, and filter,
existing active and scheduled toasts for their unique headers. This is useful
for managing state across PowerShell scripts with toasts using the same Header.
This commit also includes tests and help text for the new function and any
private helper functions.